fix django 1.11 deprecations
