+++ hawc/apps/assessment/urls.py
@@ -83,7 +83,7 @@ urlpatterns = [
         name="clean_study_metrics",
     ),
     # api views
-    url(r"^api/", include(router.urls, namespace="api")),
+    url(r"^api/", include((router.urls, "api"))),
 ]
 
 admin.autodiscover()
