+++ hawc/apps/bmd/urls.py
@@ -8,7 +8,7 @@ router.register(r"session", api.Session, basename="session")
 
 app_name = "bmd"
 urlpatterns = [
-    url(r"^api/", include(router.urls, namespace="api")),
+    url(r"^api/", include((router.urls, "api"))),
     # BMD assessment-level settings
     url(
         r"^assessment/(?P<pk>\d+)/settings/$",
