+++ hawc/main/settings/base.py
@@ -59,7 +59,7 @@ TEMPLATES = [
 
 
 # Middleware
-MIDDLEWARE_CLASSES = (
+MIDDLEWARE = (
     "django.contrib.sessions.middleware.SessionMiddleware",
     "django.middleware.common.CommonMiddleware",
     "django.middleware.csrf.CsrfViewMiddleware",
