+++ hawc/main/settings/dev.py
@@ -12,7 +12,7 @@ INSTALLED_APPS += (
     "django_extensions",
 )
 
-MIDDLEWARE_CLASSES += ("debug_toolbar.middleware.DebugToolbarMiddleware",)
+MIDDLEWARE += ("debug_toolbar.middleware.DebugToolbarMiddleware",)
 
 INTERNAL_IPS = [
     "127.0.0.1",
